python-ai/pyproject.toml
[project]
name = "agriculture-ai"
version = "1.0.0"
description = "AI models for precision agriculture platform"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "Pillow>=10.0.0",
    "scikit-learn>=1.3.0",
    "albumentations>=1.3.0",
]
